pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = "kucing mikir keras"
authors = ["Rizal Nugraha dan tama"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
"discord.py" = "^1.7.3"
revChatGPT = "^0.0.14"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"